# expression: implement vectorized evaluation for `builtinWeekWithModeSig`