--- a/expression/builtin_time_vec.go
+++ b/expression/builtin_time_vec.go
@@ -952,7 +952,7 @@ func (b *builtinWeekWithModeSig) vecEvalInt(input *chunk.Chunk, result *chunk.Co
 			continue
 		}
 		mode := int(ms[i])
-		week := date.Time.Week(mode)
+		week := date.Time.Week(int(mode))
 		i64s[i] = int64(week)
 	}
 	return nil