PCP #12103
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinWeekWithModeSig

### What is changed and how it works?
```
$ make vectorized-bench VB_FILE=Time VB_FUNC=builtinWeekWithModeSig
cd ./expression && \
        go test -v -benchmem \
                -bench=BenchmarkVectorizedBuiltinTimeFunc \
                -run=BenchmarkVectorizedBuiltinTimeFunc \
                -args "builtinWeekWithModeSig"
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-8           1000000000               0.00943 ns/op         0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinWeekWithModeSig-VecBuiltinFunc-8                 28946             41939 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinWeekWithModeSig-NonVecBuiltinFunc-8              10000            100496 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      2.765s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test